[doc] Update autorun to latest version, default to utf8 encoding
