+++ doc/sphinxext/sphinx_autorun/version.py
@@ -1,0 +1,4 @@
+# coding: utf-8
+# file generated by setuptools_scm
+# don't change, don't track in version control
+version = '1.1.1'
